+++ src/test/java/com/dukeacademy/solution/StandardProgramExecutorTest.java
@@ -4,6 +4,7 @@ import com.dukeacademy.solution.exceptions.ProgramExecutorException;
 import com.dukeacademy.solution.models.ClassFile;
 import com.dukeacademy.solution.models.ProgramInput;
 import com.dukeacademy.solution.models.ProgramOutput;
+import com.dukeacademy.solution.program.StandardProgramExecutor;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
